Fixed some property types on SeniorPlayer BusinessObject class. Fixed documentation on SortableBindingList. Added Database constants, created new SchemaName constant class and updated all mappings to include schema. Added new Mapping for SeniorPlayerWithSkillDelta view mapping and included it on model creation. Renamed and updated SeniorPlayerWithSkillDelta SQL script to use a hack to use views with EntityFramework. Adde SeniorPlayerWithSkillDelta BusinessObject class. Implemented Entity Framework migrations configuration to seed the database. Simplified some parameter declarations on FormDataFolder.

+++ Hyperar.HattrickUltimate.BusinessObjects/App/SeniorPlayer.cs
@@ -99,12 +99,12 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.App
         /// <summary>
         /// Gets or sets the Matches played on the Junior National Team.
         /// </summary>
-        public long MatchesOnJuniorNationalTeam { get; set; }
+        public short MatchesOnJuniorNationalTeam { get; set; }
 
         /// <summary>
         /// Gets or sets the Matches played on the Senior National Team.
         /// </summary>
-        public long MatchesOnSeniorNationalTeam { get; set; }
+        public short MatchesOnSeniorNationalTeam { get; set; }
 
         /// <summary>
         /// Gets or sets the NickName.
